changed test and pageobject
