+++ src/test/java/praktikum/BaseTest.java
@@ -4,23 +4,28 @@ import java.io.File;
 import java.time.Duration;
 import org.junit.After;
 import org.junit.Before;
+import org.junit.runners.Parameterized;
+import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeDriverService;
+import io.github.bonigarcia.wdm.WebDriverManager;
+
+
 
 public class BaseTest {
     protected WebDriver driver;
 
-    public BaseTest() {
+    @Before
+    public void setup() {
+        WebDriverManager.chromedriver().setup();
+        this.driver = new ChromeDriver();
     }
 
-    @Before
-    public void initDriver() {
-        System.setProperty("webdriver.http.factory", "jdk-http-client");
-        ChromeDriverService service = (ChromeDriverService)((ChromeDriverService.Builder)(new ChromeDriverService.Builder()).usingDriverExecutable(new File("C:\\webdriver\\bin\\chromedriver-win64\\chromedriver.exe"))).build();
-        this.driver = new ChromeDriver(service);
-        this.driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5L));
+    public void clickConfirmButton() {
+        this.driver.findElement(By.xpath(".//button[@id='rcc-confirm-button']")).click();
     }
+
 
     @After
     public void teardown() {
